Add "only direct"- and "only seasonal bookable" filters to units
