+++ api/graphql/types/unit/filtersets.py
@@ -1,10 +1,12 @@
 import django_filters
+from django.db import models
 from django.db.models import Count, Q
 from graphene_django_extensions import ModelFilterSet
 from graphene_django_extensions.filters import IntMultipleChoiceFilter
 
 from api.graphql.types.unit.types import Unit
 from common.date_utils import local_datetime
+from reservation_units.enums import ReservationKind
 
 __all__ = [
     "UnitFilterSet",
@@ -24,6 +26,9 @@ class UnitFilterSet(ModelFilterSet):
     published_reservation_units = django_filters.BooleanFilter(method="get_published_reservation_units")
     own_reservations = django_filters.BooleanFilter(method="get_own_reservations")
 
+    only_direct_bookable = django_filters.BooleanFilter(method="get_only_direct_bookable")
+    only_seasonal_bookable = django_filters.BooleanFilter(method="get_only_seasonal_bookable")
+
     class Meta:
         model = Unit
         fields = {
@@ -32,6 +37,7 @@ class UnitFilterSet(ModelFilterSet):
             "name_en": ["exact", "icontains", "istartswith"],
         }
         order_by = [
+            "pk",
             "name_fi",
             "name_en",
             "name_sv",
@@ -112,3 +118,23 @@ class UnitFilterSet(ModelFilterSet):
             return qs.filter(units_with_reservations)
 
         return qs.exclude(units_with_reservations)
+
+    def get_only_direct_bookable(self, qs: models.QuerySet, name: str, value: bool) -> models.QuerySet:
+        if value:
+            return qs.filter(
+                reservationunit__reservation_kind__in=[
+                    ReservationKind.DIRECT,
+                    ReservationKind.DIRECT_AND_SEASON,
+                ],
+            )
+        return qs
+
+    def get_only_seasonal_bookable(self, qs: models.QuerySet, name: str, value: bool) -> models.QuerySet:
+        if value:
+            return qs.filter(
+                reservationunit__reservation_kind__in=[
+                    ReservationKind.SEASON,
+                    ReservationKind.DIRECT_AND_SEASON,
+                ],
+            )
+        return qs
